+++ lib/src/enum/enum_orientation.dart
@@ -1,0 +1,8 @@
+enum EnumOrientation {
+  // 縦
+  vertical,
+  // 横
+  horizontal,
+  // 縦横両対応
+  any
+}