+++ lib/src/enum/spwml_enum_all.dart
@@ -1,9 +1,11 @@
 export 'enum_btn_type.dart';
 export 'enum_clip_type.dart';
+export 'enum_device_type.dart';
 export 'enum_font_style.dart';
 export 'enum_font_weight.dart';
 export 'enum_img_type.dart';
 export 'enum_indicator_type.dart';
+export 'enum_orientation.dart';
 export 'enum_spwml_element_type.dart';
 export 'enum_spwml_params.dart';
 export 'enum_text_decoration.dart';